requirements.txt
langchain==0.1.14
langchain-openai==0.1.5
langchain-community==0.0.35
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.28.1
faiss-cpu==1.13.2
PyPDF2==3.0.1
python-docx==0.8.11
tiktoken==0.5.2
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
